requirements.txt
torch>=1.10.0
torchsummary>=1.5.1
scikit-learn>=0.24.2
ignite>=0.4.10
matplotlib>=3.4.3
Pillow>=8.2.0